agents_first_project/requirements.txt
alpha-vantage
pandas
python-dotenv
scikit-learn
hmmlearn